Moved chat components into separate folder
