+++ client/src/components/chat/Cables.js
@@ -1,12 +1,12 @@
 import React, { Fragment } from 'react';
-import { ActionCable } from 'react-actioncable-provider';
+import { ActionCableConsumer } from 'react-actioncable-provider';
 
 const Cable = ({ conversations, handleReceivedMessage }) => {
   return (
     <Fragment>
       {conversations.map(conversation => {
         return (
-          <ActionCable
+          <ActionCableConsumer
             key={conversation.id}
             channel={{ channel: 'MessagesChannel', conversation: conversation.id }}
             onReceived={handleReceivedMessage}
