+++ client/src/components/chat/ConversationsList.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { ActionCable } from 'react-actioncable-provider';
-import { API_ROOT } from '../constants';
+import { ActionCableConsumer } from 'react-actioncable-provider';
+import { API_ROOT } from '../../constants';
 import NewConversationForm from './NewConversationForm';
 import MessagesArea from './MessagesArea';
 import Cable from './Cables';
@@ -42,7 +42,7 @@ class ConversationsList extends React.Component {
     const { conversations, activeConversation } = this.state;
     return (
       <div className="conversationsList">
-        <ActionCable
+        <ActionCableConsumer
           channel={{ channel: 'ConversationsChannel' }}
           onReceived={this.handleReceivedConversation}
         />
