+++ client/src/components/chat/MessagesArea.js
