+++ client/src/components/chat/NewConversationForm.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { API_ROOT, HEADERS } from '../constants';
+import { API_ROOT, HEADERS } from '../../constants';
 
 class NewConversationForm extends React.Component {
   state = {
