+++ client/src/components/chat/NewMessageForm.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { API_ROOT, HEADERS } from '../constants';
+import { API_ROOT, HEADERS } from '../../constants';
 
 class NewMessageForm extends React.Component {
   state = {
